websearch-agent-tavily/pyproject.toml
[project]
name = "websearch-agentsdk-using-tavily"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.1",
]
